requirements.txt
numpy
numba
pandas
pytest
pytest-cov
environ-config